Copy scedule logic to algorithm module
